Add Bits
This only display's Bits. You can't create them yet.

I wanted to call the model `Link` but it appears Crystal itself has
something called link: an Annotation. I'm not sure what that is yet,
but if you try to do this:

```crystal
class Link
end
```

You get an error

```
Link is not a class, it'a an annotation.
```

Searching around the crystal source code I couldn't find many capital
`L` links, but it's obviously there somewhere. I went through a couple
of other names: websites, shares, weblinks, hyperlinks, but decided on
bits. It's short, doesn't really mean anything in this context so it
just becomes a domain word.