User can request to follow other user
The idea here is that I want people to follow other people's links.
Think of it like Twitter or Instagram: One user posts stuff and other
users see it. The difference here is that you must request a follow
instead of it just automatically handling. It's like all accounts are
private by default.

This uses a FollowRequestForm to do most of the heavy lifting. A
`Follow` record needs a from_id and a to_id to show the direction of
the follow. The form requires an `email` and the `current_user`. When
created, it takes the email and sees if it can find a user that
matches. If it can, it assigns the found user to the `to_id` and the
`current_user` as the `from_id`.

The Follow still has to be accepted by the `to` `User` but this is
where I'm starting.