Accept pending follow requests
The requests come into the `/me` page. Clicking accept will set the
`accepted_at` value in the database and consider the follow accepted.

To split up follows by accepted and not accepted, we looked for the
`NULL`ness of the `accepted_at` column. If it's `NULL` it has not yet
been accepted.

We also had to grab the latest version of lucky_record because of a bug
that affected our model.
(https://github.com/luckyframework/lucky_record/pull/306).