Add Feed Token
Since feeds will be private to each user, we need a way to authenticate
and authorize the user through just a parameter in the URL. That
parameter will be this feed token.

Right now since there is no actual production (or staging) site, and
therefore no users, I can get away with filling the token with an
invalid value. During sign up, the token will be set and everything
will be fine.