Render XML feed for a user
If passed a valid token, the path /feed will render a (mostly) valid
atom xml feed for use with a feed reader. The token "parsing" is just
looking a user up by their feed_token after passing in as a parameter.
If a user is found, it sets it as current_user. Otherwise, an error is
displayed.

I created a new type of response: atom. This only sets the content type
to "application/atom+xml". I also created a small wrapper that wraps an
XML builder in the standard atom <feed> tag.

A couple of feed concessions at this point... I don't have a link to a
particular bit which prevents me from having an <id> tag, which
prevents the atom feed from being totally valid.

Similarly, I don't really have a user show page. There's /me which i'm
using as a stand in for now. Also users don't have names, so I'm just
using an email and that's technically not valid.

I'm also not yet parsing the bit descriptions into Markdown. Some of
the feed entries don't have content because there is no description.
Eventually this would be just an empty tag.

These are all minor things that are easy to solve. I felt that getting
a mostly-valid feed up was worth it before all the meta things
supporting a fully-valid feed existed.