Follow a user from their profile page
This uses a new form called FollowRequestDirectForm. It's similar to
the FollowRequestForm that takes an email (soon to be renamed
FollowRequestEmailForm) but it takes a User instead. It relies on the
fact that a new action Users::Follows::Create is a nested action and
will have a user_id parameter already present to find a user with.