Ensure usernames are case-insensitively unique
When adding a uniqueness validation to a form, you can specify a query
that the validation will use to check data in the database against the
incoming value to be saved. Without a custom query, you'll get SQL like:

```
SELECT * FROM users WHERE users.username = 'cooluser'
```

A query can transform the query. The query
`UserQuery.new.username.lower` will then produce SQL like:

```
SELECT * FROM users WHERE LOWER(users.username) = 'cooluser'
```

However, this only makes the usernames from the database lowercase. If
a user inputs a username of `COOLUSER` there will not be a match
because all of the usernames from the database were lowercased and are
trying to match against something with non-lowercase words.

The fix is to lowercase the field value before it's checked in the
database:

```
username.value = username.value.try(&.downcase)
```

which will then convert `COOLUSER` to `cooluser` and the SQL will atch
an existing user correctly.