Add bookmarklets for creating bits for each group
This adds links on the group index that a user can add to their
bookmarks bar. When clicked, the bookmarklet visits the new bit page
for that group and fills in the form. This also includes some small
style refactoring to make sections on the group index stand out more.