Add data migration for legacy link site
This migrates all bit data from the legacy site to the new one. It
requires that there is:

* A group named `LuckyBits`
* An environment variable called `LEGACY_DATABASE_URL` that is set to
the URL of the legacy database
* Two users on both systems that have matching usernames to map each
bit to

This connects to the foreign database and creates all the same bits in
a group called LuckyBits. It maps the user to one of the same username
and keeps all the timestamps, but it does not keep the bit IDs.