And empty feed does not crash the site
When a user visits a feed, the feed needs to have a last_updated date.
The date is currently calculated by the date of the most recent bit
posted to the group. If there are no bits in the group, there is no
most recent bit, and the feed would crash. Now it optionally loads the
most recent bit and uses that date, or Jan 1, 1970 if there is no bit.