Switch alternate/related link in feed
An Atom feed uses the link tag to determine what URL is linked for the
item's title. I want it to go to the thing that the user linked. For
this, it should be <link rel="alternate" ... /> but I was using
rel="related" instead which links to the group on luckybits. This
switches them.